# packages/api-contract/openapi/openapi.yaml
openapi: 3.1.1
info:
  title: KubeQueue API
  version: 1.0.0
  description: Queue and lifecycle control for standard Kubernetes batch Jobs.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /api/v1
    description: Same-origin API
security:
  - bearerAuth: []
paths:
...
  /audit/events:
    get:
      summary: Search local audit events
      description: Returns one bounded page from the customer-operated local audit record in stable occurredAt then id order. Inaccessible project scope returns an empty page.
      operationId: searchAuditEvents
      parameters:
        - { $ref: '#/components/parameters/AuditInstallationId' }
        - { $ref: '#/components/parameters/AuditProjectId' }
        - { $ref: '#/components/parameters/AuditAction' }
        - { $ref: '#/components/parameters/AuditDecision' }
        - { $ref: '#/components/parameters/AuditResult' }
        - { $ref: '#/components/parameters/AuditOccurredFrom' }
        - { $ref: '#/components/parameters/AuditOccurredTo' }
        - { $ref: '#/components/parameters/AuditLimit' }
        - { $ref: '#/components/parameters/AuditCursor' }
      responses:
        '200':
          description: Authorized audit events in deterministic keyset order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditEventPage' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /audit/events/{auditEventId}:
    parameters:
      - { $ref: '#/components/parameters/AuditEventId' }
      - { $ref: '#/components/parameters/AuditInstallationId' }
    get:
      summary: Get a local audit event
      description: Returns one event only when it exists in the caller's authorized installation and project scope; all other cases return the same not-found response.
      operationId: getAuditEvent
      responses:
        '200':
          description: Authorized local audit event
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditEvent' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
  /audit/export:
    get:
      summary: Export local audit events
      description: Returns an ordered bounded page from the local audit record for customer-owned export tooling. Follow nextCursor until null while preserving the same filters.
      operationId: exportAuditEvents
      parameters:
        - { $ref: '#/components/parameters/AuditInstallationId' }
        - { $ref: '#/components/parameters/AuditProjectId' }
        - { $ref: '#/components/parameters/AuditAction' }
        - { $ref: '#/components/parameters/AuditDecision' }
        - { $ref: '#/components/parameters/AuditResult' }
        - { $ref: '#/components/parameters/AuditOccurredFrom' }
        - { $ref: '#/components/parameters/AuditOccurredTo' }
        - { $ref: '#/components/parameters/AuditLimit' }
        - { $ref: '#/components/parameters/AuditCursor' }
      responses:
        '200':
          description: Export page in stable occurredAt then id order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditEventPage' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
  /events:
    get:
      summary: Stream job changes
      description: |
        Streams authorization-scoped job invalidations as server-sent events. Every message uses
        the `jobs` event name and a decimal event ID equal to its cursor. The first message is a
        `ready` payload. Without `Last-Event-ID`, its cursor is the latest authorized change and
        only later changes are emitted. With `Last-Event-ID`, changes after that cursor are replayed.
        An `invalidate` payload contains the unique job IDs changed in a bounded backend page;
        clients refetch affected resources and persist its cursor for reconnection.
      operationId: streamEvents
      parameters:
        - name: Last-Event-ID
          in: header
          description: Non-negative decimal cursor from the last processed SSE message.
          schema: { type: string }
      responses:
        '200':
          description: |
            SSE stream of `jobs` events. Each `data` value is either
            `{"type":"ready","cursor":0}` or
            `{"type":"invalidate","jobIds":["<job UUID>"],"cursor":1}` with current values.
          content:
            text/event-stream:
              schema: { type: string }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    sessionAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Opaque `Session <credential>` supplied only by the trusted BFF.
  headers:
...
    RequestId:
      description: Bounded request correlation identifier accepted from X-Request-ID when valid, otherwise generated by the API.
      schema: { $ref: '#/components/schemas/RequestId' }
  parameters:
...
    AuditEventId:
      name: auditEventId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/AuditIdentifier' }
    AuditInstallationId:
      name: installationId
      in: query
      required: true
      description: Customer-operated installation whose local audit record is queried.
      schema: { $ref: '#/components/schemas/AuditIdentifier' }
    AuditProjectId:
      name: projectId
      in: query
      description: Restricts results to one authorized project. Inaccessible project scope returns no events.
      schema: { $ref: '#/components/schemas/AuditIdentifier' }
    AuditAction:
      name: action
      in: query
      schema:
        {
          type: string,
          minLength: 1,
          maxLength: 96,
          pattern: '^[a-z][a-z0-9]*(?:[._-][a-z0-9]+)*$',
        }
    AuditDecision:
      name: decision
      in: query
      schema: { $ref: '#/components/schemas/AuditAuthorizationDecision' }
    AuditResult:
      name: result
      in: query
      schema: { $ref: '#/components/schemas/AuditResult' }
    AuditOccurredFrom:
      name: occurredFrom
      in: query
      description: Includes events occurring at or after this RFC 3339 timestamp.
      schema: { type: string, format: date-time }
    AuditOccurredTo:
      name: occurredTo
      in: query
      description: Includes events occurring before this RFC 3339 timestamp.
      schema: { type: string, format: date-time }
    AuditLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    AuditCursor:
      name: cursor
      in: query
      description: Opaque continuation token preserving the stable occurredAt and id keyset. It is valid only with the same filters.
      schema: { type: string, maxLength: 512 }
  responses:
    Error:
      description: Request failed
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: The resource is absent or not visible in the authorized scope.
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            requestId: 550e8400-e29b-41d4-a716-446655440000
            error:
              { code: RESOURCE_NOT_FOUND, message: Resource not found, status: 404, details: {} }
...
  schemas:
    RequestId:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9][A-Za-z0-9._:@/-]*$'
...
    AuditIdentifier:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9][A-Za-z0-9._:@/-]*$'
    AuditAuthorizationDecision:
      type: string
      enum: [ALLOW, DENY]
    AuditResult:
      type: string
      enum: [SUCCESS, FAILURE]
    AuditAuthenticationMethod:
      type: string
      enum:
        [OIDC_CLIENT_CREDENTIALS, OIDC_SESSION, SERVICE_ACCOUNT_TOKEN, BREAK_GLASS, LEGACY_TOKEN]
    AuditSourceProvenance:
      type: string
      enum: [DIRECT_PEER, TRUSTED_PROXY]
    AuditActor:
      type: object
      additionalProperties: false
      required: [principalId, authenticationMethod, credentialId, effectiveGroups]
      properties:
        principalId: { $ref: '#/components/schemas/AuditIdentifier' }
        authenticationMethod: { $ref: '#/components/schemas/AuditAuthenticationMethod' }
        credentialId: { $ref: '#/components/schemas/AuditIdentifier' }
        effectiveGroups:
          type: array
          maxItems: 64
          uniqueItems: true
          items: { $ref: '#/components/schemas/AuditIdentifier' }
    AuditTarget:
      type: object
      additionalProperties: false
      required: [type, id]
      properties:
        type: { type: string, minLength: 1, maxLength: 64 }
        id: { $ref: '#/components/schemas/AuditIdentifier' }
    AuditScope:
      type: object
      additionalProperties: false
      required: [installationId]
      properties:
        installationId: { $ref: '#/components/schemas/AuditIdentifier' }
        projectId: { $ref: '#/components/schemas/AuditIdentifier' }
        teamId: { $ref: '#/components/schemas/AuditIdentifier' }
        namespace: { type: string, minLength: 1, maxLength: 63 }
    AuditSource:
      type: object
      additionalProperties: false
      required: [address, provenance, userAgent]
      properties:
        address: { type: string, minLength: 2, maxLength: 45 }
        provenance: { $ref: '#/components/schemas/AuditSourceProvenance' }
        userAgent: { type: string, maxLength: 512 }
    AuditSummary:
      type: object
      additionalProperties: false
      required: [changedFields, redactionCount, truncated]
      properties:
        state: { type: string, maxLength: 64, pattern: '^[A-Z][A-Z0-9_]{0,63}$' }
        changedFields:
          type: array
          maxItems: 64
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 128 }
        redactionCount: { type: integer, minimum: 0, maximum: 65535 }
        truncated: { type: boolean }
    AuditEvent:
      type: object
      additionalProperties: false
      required:
        [
          id,
          occurredAt,
          requestId,
          traceId,
          actor,
          action,
          target,
          scope,
          decision,
          result,
          reason,
          source,
        ]
      properties:
        id: { $ref: '#/components/schemas/AuditIdentifier' }
        occurredAt: { type: string, format: date-time }
        requestId: { $ref: '#/components/schemas/AuditIdentifier' }
        traceId: { $ref: '#/components/schemas/AuditIdentifier' }
        actor: { $ref: '#/components/schemas/AuditActor' }
        action: { type: string, minLength: 1, maxLength: 96 }
        target: { $ref: '#/components/schemas/AuditTarget' }
        scope: { $ref: '#/components/schemas/AuditScope' }
        decision: { $ref: '#/components/schemas/AuditAuthorizationDecision' }
        result: { $ref: '#/components/schemas/AuditResult' }
        reason: { type: string, minLength: 1, maxLength: 96 }
        source: { $ref: '#/components/schemas/AuditSource' }
        before: { $ref: '#/components/schemas/AuditSummary' }
        after: { $ref: '#/components/schemas/AuditSummary' }
    AuditEventPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          maxItems: 200
          items: { $ref: '#/components/schemas/AuditEvent' }
        nextCursor:
          type: [string, 'null']
          maxLength: 512
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [requestId, error]
      properties:
        requestId: { $ref: '#/components/schemas/RequestId' }
        error:
          type: object
          additionalProperties: false
          required: [code, message, status, details]
          properties:
            code: { type: string, minLength: 1, maxLength: 96 }
            message: { type: string, minLength: 1, maxLength: 512 }
            status: { type: integer, minimum: 400, maximum: 599 }
            details:
              type: object
              maxProperties: 16
              propertyNames:
                pattern: '^[A-Za-z][A-Za-z0-9_.-]{0,63}$'
              additionalProperties:
                type: string
                maxLength: 512